Worked on Scala Slick not being able to convert Boolean to Bit in postgres, changed some models, and added new create and update response to track status of those operations.

+++ app/DAL/Models/UserDetail.scala
@@ -2,4 +2,4 @@ package DAL.Models
 
 case class UserDetail (id: Long, description: String, country: String, religion:String,
                        height: Double, weight: Double, skin: String, hair: String, gender: String,
-                       birthYear: Int, birthMonth: Int, birthDay: Int, userId: Long)
+                       birthYear: Short, birthMonth: Short, birthDay: Short, userId: Long)
